Summary Files Created
Created the the following files:

Description:
To store a description of the reason for this experiment, the game and its rules, the reward function, the AIs and their training method, and our method of evaluation.

Hypothesis:
To store our hypothesis about the outcome of this experiment. This includes all different possibilities of behavior which the AIs could converge to, and our guess at which is most likely, and why.

Results:
To store the results of our outcome. This should include a description of which behavior was converged on, and how the AI's performed during evaluation.

Model:
To store a description of the model and how it works, why we chose this model, other models we considered, and resources we used to help learn about the model.

Objectives:
To store a list of objectives to be completed during each week, and a list of topics to discuss and work on during meetings. Additionally, we should record meeting notes and what we completed each week.